use hook for widgets
